elevator/requirement.txt
asgiref==3.6.0
Django==3.2.6
django-cors-headers==3.13.0
djangorestframework==3.14.0
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
tendo==0.3.0